templates/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
openstacksdk==1.3.0
mysql-connector-python==8.2.0 